+++ foxtive/CHANGELOG.md
@@ -1,2 +1,7 @@
 # Foxtive Changelog
 Foxtive changelog file 
+
+### 0.4.0
+* Added database shareable ext
+* Move database traits to 'ext' folder
+* impl From<crate::Error> for AppMessage